requirements.txt
oauthlib
flask
pyopenssl
cryptography
pyjwt
requests